Add edit User and making first steps to generic Table
