+++ services/web-ui/.eslintrc.yaml
@@ -15,6 +15,7 @@ extends:
 - plugin:jsx-a11y/recommended
 rules:
   max-len: 0
+  no-underscore-dangle: 0
   class-methods-use-this: 0
   no-console: 0
   import/prefer-default-export: 0
