--- services/web-ui/src/component/user-table/index.js
@@ -1,383 +1,0 @@
-import React from 'react';
-import { withStyles } from '@material-ui/styles';
-import Grid from '@material-ui/core/Grid';
-import flow from 'lodash/flow';
-import classNames from 'classnames';
-import PropTypes from 'prop-types';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableHead from '@material-ui/core/TableHead';
-import TablePagination from '@material-ui/core/TablePagination';
-import TableRow from '@material-ui/core/TableRow';
-import TableSortLabel from '@material-ui/core/TableSortLabel';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import Paper from '@material-ui/core/Paper';
-import Checkbox from '@material-ui/core/Checkbox';
-import IconButton from '@material-ui/core/IconButton';
-import Tooltip from '@material-ui/core/Tooltip';
-import DeleteIcon from '@material-ui/icons/Delete';
-import FilterListIcon from '@material-ui/icons/FilterList';
-import Input from '@material-ui/core/Input';
-
-import Loader from '../loader';
-
-let counter = 0;
-
-const rows = [
-    {
-        id: 'username', numeric: false, disablePadding: true, label: 'Username',
-    },
-    {
-        id: 'role', numeric: true, disablePadding: false, label: 'Role',
-    },
-    {
-        id: 'created', numeric: true, disablePadding: false, label: 'Creaded At',
-    },
-    {
-        id: 'status', numeric: true, disablePadding: false, label: 'Status',
-    },
-    {
-        id: 'actions', numeric: true, disablePadding: false, label: 'Actions',
-    },
-];
-
-function createData(username, role, created, status, actions) {
-    counter += 1;
-    return {
-        id: counter, username, role, created, status, actions,
-    };
-}
-function desc(a, b, orderBy) {
-    if (b[orderBy] < a[orderBy]) {
-        return -1;
-    }
-    if (b[orderBy] > a[orderBy]) {
-        return 1;
-    }
-    return 0;
-}
-
-function stableSort(array, cmp) {
-    const stabilizedThis = array.map((el, index) => [el, index]);
-    stabilizedThis.sort((a, b) => {
-        const order = cmp(a[0], b[0]);
-        if (order !== 0) return order;
-        return a[1] - b[1];
-    });
-    return stabilizedThis.map(el => el[0]);
-}
-
-function getSorting(order, orderBy) {
-    return order === 'desc' ? (a, b) => desc(a, b, orderBy) : (a, b) => -desc(a, b, orderBy);
-}
-
-const toolbarStyles = () => ({
-    root: {
-        paddingRight: 10,
-    },
-    highlight: {
-        color: 'secondary',
-    },
-    spacer: {
-        flex: '1 1 100%',
-    },
-    actions: {
-        flex: '0 0 auto',
-        color: 'secondary',
-    },
-    title: {
-        flex: '0 0 auto',
-    },
-});
-
-let TableToolbar = (props) => {
-    const { numSelected, classes } = props;
-
-    return (
-        <Toolbar
-            className={classNames(classes.root, {
-                [classes.highlight]: numSelected > 0,
-            })}
-        >
-            <div className={classes.title}>
-                {numSelected > 0 ? (
-                    <Typography color="inherit" variant="h6">
-                        {numSelected} Users selected
-                    </Typography>
-                ) : (
-                    <Typography variant="h6" id="tableTitle">
-              Users
-                    </Typography>
-                )}
-            </div>
-            <div className={classes.spacer} />
-            <div className={classes.actions}>
-                <Input id="filter" name="filter" value={props.filter} />
-                <Tooltip title="Filter list">
-                    <IconButton aria-label="Filter list">
-                        <FilterListIcon />
-                    </IconButton>
-                </Tooltip>
-                <Tooltip title="Delete">
-                    <IconButton aria-label="Delete">
-                        <DeleteIcon />
-                    </IconButton>
-                </Tooltip>
-            </div>
-        </Toolbar>
-    );
-};
-
-TableToolbar.propTypes = {
-    classes: PropTypes.object.isRequired,
-    numSelected: PropTypes.number.isRequired,
-};
-
-TableToolbar = withStyles(toolbarStyles)(TableToolbar);
-
-
-class TableHeader extends React.Component {
-    createSortHandler = property => (event) => {
-        this.props.onRequestSort(event, property);
-    };
-
-    render() {
-        const {
-            onSelectAllClick, order, orderBy, numSelected, rowCount,
-        } = this.props;
-        return (
-
-            <TableHead>
-                <TableRow>
-                    <TableCell padding="checkbox">
-                        <Checkbox
-                            indeterminate={numSelected > 0 && numSelected < rowCount}
-                            checked={numSelected === rowCount}
-                            onChange={onSelectAllClick}
-                        />
-                    </TableCell>
-                    {rows.map(
-                        row => (
-                            <TableCell
-                                key={row.id}
-                                align={row.numeric ? 'right' : 'left'}
-                                padding={row.disablePadding ? 'none' : 'default'}
-                                sortDirection={orderBy === row.id ? order : false}
-                            >
-                                <Tooltip
-                                    title="Sort"
-                                    placement={row.numeric ? 'bottom-end' : 'bottom-start'}
-                                    enterDelay={300}
-                                >
-                                    <TableSortLabel
-                                        active={orderBy === row.id}
-                                        direction={order}
-                                        onClick={this.createSortHandler(row.id)}
-                                    >
-                                        {row.label}
-                                    </TableSortLabel>
-                                </Tooltip>
-                            </TableCell>
-                        ),
-                        this,
-                    )}
-                </TableRow>
-            </TableHead>
-        );
-    }
-}
-
-TableHeader.propTypes = {
-    numSelected: PropTypes.number.isRequired,
-    onRequestSort: PropTypes.func.isRequired,
-    onSelectAllClick: PropTypes.func.isRequired,
-    order: PropTypes.string.isRequired,
-    orderBy: PropTypes.string.isRequired,
-    rowCount: PropTypes.number.isRequired,
-};
-
-const useStyles = () => ({
-    wrapper: {
-        width: '100%',
-        padding: '10vh 0 0 0',
-    },
-    root: {
-        width: '100%',
-        marginTop: 10,
-    },
-    table: {
-        minWidth: 1020,
-    },
-    tableWrapper: {
-        overflowX: 'auto',
-    },
-});
-class UserTable extends React.Component {
-    state = {
-        order: 'asc',
-        orderBy: 'username',
-        selected: [],
-        page: 0,
-        data: [],
-        rowsPerPage: 10,
-    };
-
-    componentDidUpdate(prefProps) {
-        const tempArr = [];
-        if (prefProps.users !== this.props.users && this.props.users && this.props.users.length > 0) {
-            this.props.users.forEach((user) => {
-                tempArr.push(createData(user.username, user.role, user.createdAt, user.status, 'actions'));
-            });
-            this.setState({
-                data: tempArr,
-            });
-        }
-    }
-
-    handleRequestSort = (event, property) => {
-        const orderBy = property;
-        let order = 'desc';
-
-        if (this.state.orderBy === property && this.state.order === 'desc') {
-            order = 'asc';
-        }
-
-        this.setState({ order, orderBy });
-    };
-
-    handleSelectAllClick = (event) => {
-        if (event.target.checked) {
-            this.setState(state => ({ selected: state.data.map(n => n.id) }));
-            return;
-        }
-        this.setState({ selected: [] });
-    };
-
-    handleClick = (event, id) => {
-        const { selected } = this.state;
-        const selectedIndex = selected.indexOf(id);
-        let newSelected = [];
-
-        if (selectedIndex === -1) {
-            newSelected = newSelected.concat(selected, id);
-        } else if (selectedIndex === 0) {
-            newSelected = newSelected.concat(selected.slice(1));
-        } else if (selectedIndex === selected.length - 1) {
-            newSelected = newSelected.concat(selected.slice(0, -1));
-        } else if (selectedIndex > 0) {
-            newSelected = newSelected.concat(
-                selected.slice(0, selectedIndex),
-                selected.slice(selectedIndex + 1),
-            );
-        }
-
-        this.setState({ selected: newSelected });
-    };
-
-    handleChangePage = (event, page) => {
-        this.setState({ page });
-    };
-
-    handleChangeRowsPerPage = (event) => {
-        this.setState({ rowsPerPage: event.target.value });
-    };
-
-    isSelected = id => this.state.selected.indexOf(id) !== -1;
-
-    render() {
-        const { classes } = this.props;
-        let emptyRows;
-        const {
-            data, order, orderBy, selected, rowsPerPage, page,
-        } = this.state;
-        if (data) {
-            emptyRows = rowsPerPage - Math.min(rowsPerPage, data.length - page * rowsPerPage);
-            return (
-                <div className={classes.wrapper}>
-                    <Grid container >
-                        <Grid item xs={12}>
-                            <Paper className={classes.root}>
-                                <TableToolbar numSelected={selected.length} setFilter={this.handleRequestSort} filter={this.state.orderBy}/>
-                                <div className={classes.tableWrapper}>
-                                    <Table className={classes.table} aria-labelledby="tableTitle">
-                                        <TableHeader
-                                            numSelected={selected.length}
-                                            order={order}
-                                            orderBy={orderBy}
-                                            onSelectAllClick={this.handleSelectAllClick}
-                                            onRequestSort={this.handleRequestSort}
-                                            rowCount={data.length}
-                                        />
-                                        <TableBody>
-                                            {stableSort(data, getSorting(order, orderBy))
-                                                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                                                .map((n) => {
-                                                    const isSelected = this.isSelected(n.id);
-                                                    return (
-                                                        <TableRow
-                                                            hover
-                                                            onClick={event => this.handleClick(event, n.id)}
-                                                            role="checkbox"
-                                                            aria-checked={isSelected}
-                                                            tabIndex={-1}
-                                                            key={n.id}
-                                                            selected={isSelected}
-                                                        >
-                                                            <TableCell padding="checkbox">
-                                                                <Checkbox checked={isSelected} />
-                                                            </TableCell>
-                                                            <TableCell component="th" scope="row" padding="none">
-                                                                {n.username}
-                                                            </TableCell>
-                                                            <TableCell align="right">{n.role}</TableCell>
-                                                            <TableCell align="right">{n.created}</TableCell>
-                                                            <TableCell align="right">{n.status}</TableCell>
-                                                            <TableCell align="right">{n.actions}</TableCell>
-                                                        </TableRow>
-                                                    );
-                                                })}
-                                            {emptyRows > 0 && (
-                                                <TableRow style={{ height: 49 * emptyRows }}>
-                                                    <TableCell colSpan={6} />
-                                                </TableRow>
-                                            )}
-                                        </TableBody>
-                                    </Table>
-                                </div>
-                                <TablePagination
-                                    rowsPerPageOptions={[5, 10, 25]}
-                                    component="div"
-                                    count={data.length}
-                                    rowsPerPage={rowsPerPage}
-                                    page={page}
-                                    backIconButtonProps={{
-                                        'aria-label': 'Previous Page',
-                                    }}
-                                    nextIconButtonProps={{
-                                        'aria-label': 'Next Page',
-                                    }}
-                                    onChangePage={this.handleChangePage}
-                                    onChangeRowsPerPage={this.handleChangeRowsPerPage}
-                                />
-                            </Paper>
-                        </Grid>
-                    </Grid>
-                </div>
-            );
-        }
-
-        return (
-            <Loader/>
-        );
-    }
-}
-
-UserTable.propTypes = {
-    classes: PropTypes.object.isRequired,
-};
-
-export default flow(
-    withStyles(useStyles),
-)(UserTable);
