+++ services/web-ui/src/conf/index.js
@@ -1,4 +1,18 @@
-let config = {};
+let config = {
+    tableConfig: {
+        user: {
+            username: 'Username',
+            role: 'Role',
+            createdAt: 'Created At',
+            status: 'Status',
+        },
+        tenant: {
+            tenantname: 'Organisations',
+            createdAt: 'Created At',
+            status: 'Status',
+        },
+    },
+};
 
 export function getConfig() {
     return config;
